LayoutTests/imported/w3c/web-platform-tests/webcodecs/audio-encoder.https.any-expected.txt

CONSOLE MESSAGE: Error: assert_unreached: error: NotSupportedError: Not supported Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: error: NotSupportedError: Not supported Reached unreachable code
CONSOLE MESSAGE: NotSupportedError: AudioData creation failed

Harness Error (FAIL), message = NotSupportedError: AudioData creation failed

FAIL Simple audio encoding promise_test: Unhandled rejection with value: object "NotSupportedError: AudioData creation failed"
FAIL Test reset during flush assert_unreached: unexpected error Reached unreachable code
FAIL Encode audio with negative timestamp promise_test: Unhandled rejection with value: object "NotSupportedError: AudioData creation failed"
